#MARK-31 Handling Failed AddOverlays()

+++ app/Services/Inventory/InventoryService.php
@@ -656,8 +656,6 @@ class InventoryService implements InventoryServiceInterface
 
         $overlayParams = $this->inventoryRepository->getOverlayParams($inventoryId);
 
-        Log::channel('inventory-overlays')->info('Adding Overlays on Inventory Images', $overlayParams);
-
         $overlayEnabled = $overlayParams['dealer_overlay_enabled'] ?? $overlayParams['overlay_enabled'];
 
         foreach ($inventoryImages as $inventoryImage) {
@@ -676,13 +674,21 @@ class InventoryService implements InventoryServiceInterface
                 $originalFilename = !empty($imageObj->filename_noverlay) ? $imageObj->filename_noverlay : $imageObj->filename;
                 $localNewImagePath = $this->imageService->addOverlays($this->getS3BaseUrl() . $originalFilename, $overlayParams);
 
-                // upload overlay image
-                $randomFilename = md5($localNewImagePath);
-                $newFilename = $this->imageService->uploadToS3($localNewImagePath, $randomFilename, $overlayParams['dealer_id']);
-                unlink($localNewImagePath);
+                if (!empty($localNewImagePath)) {
 
-                // update image to database
-                $this->imageTableService->saveOverlay($imageObj, $newFilename);
+                    // upload overlay image
+                    $randomFilename = md5($localNewImagePath);
+                    $newFilename = $this->imageService->uploadToS3($localNewImagePath, $randomFilename, $overlayParams['dealer_id']);
+                    unlink($localNewImagePath);
+    
+                    // update image to database
+                    $this->imageTableService->saveOverlay($imageObj, $newFilename);
+                } else {
+
+                    Log::channel('inventory-overlays')
+                        ->info('Failed Adding Overlays, Invalid OverlayParams', 
+                            array_merge($overlayParams, ['image_id' => $imageObj->image_id]));
+                }
 
             // otherwise Reset Overlay
             } else {
